+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Cover from './resume/Cover';
+import Cover from './components/Cover';
 import './App.css';
 
 function App() {
